ingester/requirements.txt
kafka-python==2.0.2
influxdb-client[ciso]
dataclasses-json
